remove batch_to_device change
