+++ pytorch_lightning/loops/epoch/prediction_epoch_loop.py
@@ -92,7 +92,7 @@ class PredictionEpochLoop(Loop):
             raise StopIteration
 
         with self.trainer.profiler.profile("predict_batch_to_device"):
-            batch = self.trainer.training_type_plugin.batch_to_device(batch, dataloader_idx=dataloader_idx)
+            batch = self.trainer.accelerator.batch_to_device(batch, dataloader_idx=dataloader_idx)
 
         self.batch_progress.increment_ready()
