+++ pytorch_lightning/loops/epoch/training_epoch_loop.py
@@ -157,7 +157,7 @@ class TrainingEpochLoop(loops.Loop[_OUTPUTS_TYPE]):
 
         if not self.trainer._data_connector.train_data_fetcher.store_on_device:
             with self.trainer.profiler.profile("training_batch_to_device"):
-                batch = self.trainer.training_type_plugin.batch_to_device(batch)
+                batch = self.trainer.accelerator.batch_to_device(batch)
 
         self.batch_progress.increment_ready()
